+++ src/extension.ts
@@ -3,6 +3,7 @@
 import * as vscode from 'vscode';
 import * as sortLines from './sort_lines';
 import GroovyDocumentSymbolProvider from './groovy_document_symbol_provider';
+import { TagLibIndex } from './gsp/taglib_index';
 
 export function activate(context: vscode.ExtensionContext) {
     
@@ -17,7 +18,10 @@ export function activate(context: vscode.ExtensionContext) {
 
     commands.forEach(command => context.subscriptions.push(command));
     languages.forEach(language => context.subscriptions.push(language));
-      
+
+    const tagLibIndex = new TagLibIndex();
+    context.subscriptions.push(tagLibIndex);
+    void tagLibIndex.start(context);
 }
 
 export function deactivate() {